Add files via upload
This is an interactive balance sheet. It is completely text based and can be run on python. It asks you for fields such as assets, liabilities, equity, and expenses. Once you give it, it generates a balance sheet with the debit and credit side as well as the balance, indicating if it was a credit or debit balance